requirements.txt
asgiref==3.9.1
attrs==25.3.0
black==25.1.0
click==8.2.1
Django==5.1.1
django-js-asset==3.1.2
django-mptt==0.16.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-nested-routers==0.94.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.8
psycopg==3.2.9
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
uritemplate==4.2.0
